+++ source/App/documents/release-notes/release-notes.md
@@ -1,5 +1,10 @@
 # App Release notes
 
+## Version 4.0.0
+
+- Update to the namespace for the `FunctionTelemetryScope` middleware.
+- CorrelationIdMiddleware now exposes `CorrelationId` from either a `HTTP-header` or a `ServicebusMessage` user property 
+
 ## Version 3.0.1
 
 - Bumped patch version as pipeline file was updated.